Refactor: reorganize attractor-related code
