+++ src/bang/__init__.py
@@ -1,7 +1,7 @@
 from numba import config
 
 from .core import PBN, load_from_file
-from .graph.graph import get_blocks
+from .core.attractors.blocks.graph import get_blocks
 
 config.CUDA_ENABLE_PYNVJITLINK = 1  # type: ignore
