+++ src/bang/core/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from bang.core.PBN import PBN, load_from_file
+from bang.core.pbn.pbn import PBN, load_from_file
 
 __all__ = ["PBN", "load_from_file"]
