--- src/bang/graph/__init__.py
@@ -1,3 +1,0 @@
-from bang.graph.graph import get_blocks
-
-__all__ = ["get_blocks"]
