--- src/bang/graph/example.assa
@@ -1,35 +1,0 @@
-type=synchronous
-n=6
-perturbation=0.0001
-nodeNames
-x1
-x2
-x3
-x4
-x5
-x6
-endNodeNames
-node x1
-1.0: !(x1 & x2)
-endNode
-node x2
-1.0: x1 & (!x2)
-endNode
-node x3
-1.0: !x2
-endNode
-node x4
-1.0: (x2&x3) | x5
-endNode
-node x5
-1.0: x4 | x5
-endNode
-node x6
-1.0: (!x3) & x6
-endNode
-npNode
-x1
-x2
-x4
-x5
-endNpNode
